requirements.txt
pytesseract
pillow